terraform/functions/weather/requirements.txt
functions-framework==3.*
requests
google-cloud-storage
pandas
pyarrow
fastparquet